rename[#3]: music 패키지 ai로 리네이밍
